requirements.txt
spacy==3.8.3
sumy==0.11.0
textblob==0.18.0.post0
tqdm==4.67.1
feedparser==6.0.11
pygooglenews==0.1.2
googlenewsdecoder==0.1.6
newspaper3k==0.2.8
nltk==3.9.1